txt2img/app/requirements.txt
fastapi
uvicorn 
torch 
torchvision
transformers
streamlit
pillow
python-multipart
diffusers
accelerate